# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Seatsurfing REST API
  description: |
    Seatsurfing provides REST APIs used by the mobile and web app as well as the administrator web interface.

    ## Authorization checks

    Each API endpoint is protected by authorization checks:
    1. **Permission check**: the permissions granted by the roles assigned to the calling user
    2. **Organization check**: organization membership - no request reaches data belonging to another organization

    ## Permissions

    Access is granted by roles, which an organization's administrators define. A role grants each
    functionality at a level, and a user may hold several roles, holding the highest level any of
    them grants.

    Levels: `0` no access, `10` read only, `20` limited, `30` full access. Not every level is
    meaningful for every permission; `GET /role/permissions` reports the levels each one offers.

    Permission keys: `areas`, `space_attributes`, `bookings`, `approvals`, `analytics`,
    `presence_report`, `users`, `groups`, `roles`, `service_accounts`, `auth_providers`,
    `org_settings`, `audit_log`. Plugins contribute further keys prefixed with `plugin.`.

    `analytics` covers aggregate occupancy figures; `presence_report` covers who was present
    and when. They are separate because the latter is personal data, and each is additionally
    subject to an organization-wide setting (`hide_stats` and `hide_reports`) that withholds it
...

...

...

    ## Authentication

...

...

...

...

...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained via login or refresh endpoints
    BasicAuth:
      type: http
      scheme: basic
      description: Service account authentication using `{orgId}_{email}` as username and the service account password
    ServiceAccountBearerAuth:
      type: http
      scheme: bearer
      description: Service account authentication using a long-lived API token generated via `POST /user/{id}/api-token`. The raw token is presented as a Bearer token. Read-only service accounts may only use this on `GET` requests.
  schemas:
...
    CreateLocationRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        maxConcurrentBookings:
          type: integer
        timezone:
          type: string
          description: IANA timezone (e.g. "Europe/Berlin")
        enabled:
          type: boolean
        mapScale:
          type: number
          format: double
    GetLocationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        maxConcurrentBookings:
          type: integer
        timezone:
          type: string
        enabled:
          type: boolean
        mapScale:
          type: number
          format: double
        mapWidth:
          type: integer
        mapHeight:
          type: integer
        mapMimeType:
          type: string
...
    SetSpaceAttributeValueRequest:
      type: object
      properties:
        value:
          type: string
...
  responses:
    Created:
      description: Object created successfully
      headers:
        X-Object-ID:
          schema:
            type: string
          description: ID of the created object
    Updated:
      description: Object updated or deleted successfully (204 No Content)
    BadRequest:
      description: Malformed request or incomplete JSON
      headers:
        X-Error-Code:
          schema:
            type: integer
          description: Custom error code (see error codes table)
...
    Forbidden:
      description: Authorized, but user has no access to requested resource
    NotFound:
      description: Object not found
...
paths:
...
  /location/:
    get:
      tags: [Locations]
      summary: Get all locations
      description: Returns all locations in the user's organization.
      operationId: getAllLocations
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetLocationResponse"
    post:
      tags: [Locations]
      summary: Create a location
      description: Creates a new location in the user's organization. Requires `areas` at full access.
      operationId: createLocation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /location/search:
    post:
      tags: [Locations]
      summary: Search locations
      description: Searches locations with optional attribute filters (e.g. number of free spaces, buddies on site). If no attributes are provided, returns all locations.
      operationId: searchLocations
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchLocationRequest"
      responses:
        "200":
          description: Matching locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetLocationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /location/loadsampledata:
    post:
      tags: [Locations]
      summary: Load sample data
      description: Creates sample locations and spaces for the organization. Requires `areas` at full access.
      operationId: loadSampleData
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Sample data loaded
        "403":
          $ref: "#/components/responses/Forbidden"
  /location/{id}:
    get:
      tags: [Locations]
      summary: Get a location
      description: Returns a single location by ID.
      operationId: getLocation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Location details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetLocationResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Locations]
      summary: Update a location
      description: Updates an existing location. Requires `areas` at full access.
      operationId: updateLocation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationRequest"
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Locations]
      summary: Delete a location
      description: Deletes a location and all its spaces. Requires `areas` at full access.
      operationId: deleteLocation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /location/{id}/map:
    get:
      tags: [Locations]
      summary: Get location map
      description: Returns the floor plan / map image for a location as base64-encoded data.
      operationId: getLocationMap
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Map data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMapResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Locations]
      summary: Upload location map
      description: Uploads or replaces the floor plan / map image for a location. Accepts PNG, JPEG, GIF, or SVG. Requires `areas` at full access.
      operationId: setLocationMap
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
          image/jpeg:
            schema:
              type: string
              format: binary
          image/gif:
            schema:
              type: string
              format: binary
          image/svg+xml:
            schema:
              type: string
              format: binary
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /location/{id}/attribute:
    get:
      tags: [Locations]
      summary: Get location attribute values
      description: Returns all attribute values set on a location.
      operationId: getLocationAttributes
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Location attribute values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetSpaceAttributeValueResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /location/{id}/attribute/{attributeId}:
    post:
      tags: [Locations]
      summary: Set location attribute value
      description: Sets an attribute value on a location. Only attributes marked as `locationApplicable` can be set. Requires `areas` at full access.
      operationId: setLocationAttribute
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: attributeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetSpaceAttributeValueRequest"
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Locations]
      summary: Delete location attribute value
      description: Removes an attribute value from a location. Requires `areas` at full access.
      operationId: deleteLocationAttribute
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: attributeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===========================
  # Spaces
  # ===========================
...
  /search/:
    get:
      tags: [Search]
      summary: Search across entities
      description: |
        Searches across users, locations, spaces, and groups by keyword. Requires `users` at read level or above.
        Use query parameters to control which entity types to include in the search:
        - `includeUsers=1` (requires `users` at read level or above)
        - `includeGroups=1`
        - `includeLocations=1`
        - `includeSpaces=1`
        To expand the location object (in the spaces result), use the expand query parameter:
        - `expandLocations=1`
      operationId: search
      security:
        - BearerAuth: []
      parameters:
        - name: query
          in: query
          schema:
            type: string
          description: Search keyword
        - name: includeUsers
          in: query
          schema:
            type: string
            enum: ["0", "1"]
          description: Include users in search results
        - name: includeGroups
          in: query
          schema:
            type: string
            enum: ["0", "1"]
          description: Include groups in search results
        - name: includeLocations
          in: query
          schema:
            type: string
            enum: ["0", "1"]
          description: Include locations in search results
        - name: includeSpaces
          in: query
          schema:
            type: string
            enum: ["0", "1"]
          description: Include spaces in search results
        - name: expandLocations
          in: query
          schema:
            type: string
            enum: ["0", "1"]
          description: Expand the location object in search results
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSearchResultsResponse"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ===========================
  # Confluence
  # ===========================
...